chore: move user associated models into new app, nuke all migrations
