--- backend/custom_user/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-from .models import CustomUser
-
-admin.site.register(CustomUser)
