--- backend/custom_user/apps.py
@@ -1,9 +1,0 @@
-from django.apps import AppConfig
-
-
-class CustomUserConfig(AppConfig):
-	default_auto_field = 'django.db.models.BigAutoField'
-	name = 'custom_user'
-
-	def ready(self):
-		import custom_user.signal  # noqa
