+++ backend/transcendence/settings.py
@@ -48,7 +48,7 @@ INSTALLED_APPS = [
 	'crispy_bootstrap4',
 	'django.contrib.admin',
 	'django.contrib.auth',
-	'custom_user.apps.CustomUserConfig',
+	'user_management.apps.UserManagementConfig',
 	'django.contrib.contenttypes',
 	'django.contrib.sessions',
 	'django.contrib.messages',
@@ -181,7 +181,7 @@ LOGIN_REDIRECT_URL = 'users:list'
 
 LOGIN_URL = 'users:login'
 
-AUTH_USER_MODEL = 'custom_user.CustomUser'
+AUTH_USER_MODEL = 'user_management.CustomUser'
 
 ASGI_APPLICATION = 'transcendence.asgi.application'
