+++ backend/user_management/admin.py
@@ -1,1 +1,5 @@
-# Register your models here.
+from django.contrib import admin
+
+from .models import CustomUser
+
+admin.site.register(CustomUser)
