+++ backend/user_management/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class UserManagementConfig(AppConfig):
 	default_auto_field = 'django.db.models.BigAutoField'
 	name = 'user_management'
+
+	def ready(self):
+		from . import signal  # noqa
